+++ src/main/java/org/cookieandkakao/babting/domain/calendar/service/TalkCalendarService.java
@@ -23,38 +23,24 @@ import org.springframework.web.client.RestClient;
 @Service
 public class TalkCalendarService {
 
-    private final RestClient restClient = RestClient.builder().build();
+    private final TalkCalendarClientService talkCalendarClientService;
     private final ObjectMapper objectMapper = new ObjectMapper();
     private final EventService eventService;
 
-    public TalkCalendarService(EventService eventService) {
+    public TalkCalendarService(EventService eventService, TalkCalendarClientService talkCalendarClientService) {
         this.eventService = eventService;
-    }
-
-    public EventListGetResponse getEventList(String accessToken, String from, String to) {
-        String url = "https://kapi.kakao.com/v2/api/calendar/events";
-        URI uri = buildUri(url, from, to);
-        try {
-            ResponseEntity<EventListGetResponse> response = restClient.get()
-                .uri(uri)
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .retrieve()
-                .toEntity(EventListGetResponse.class);
-            return response.getBody();
-        } catch (Exception e) {
-            throw new RuntimeException("API 호출 중 오류 발생");
-        }
+        this.talkCalendarClientService = talkCalendarClientService;
     }
 
     // 일정 목록을 조회할 때 캐시 적용
     @Cacheable(value = "eventListCache", key = "#memberId")
     public List<EventGetResponse> getUpdatedEventList(String accessToken, String from, String to, Long memberId) {
-        EventListGetResponse eventList = getEventList(accessToken, from, to);
+        EventListGetResponse eventList = talkCalendarClientService.getEventList(accessToken, from, to);
         List<EventGetResponse> updatedEvents = new ArrayList<>();
 
         for (EventGetResponse event : eventList.events()) {
             if (event.id() != null) {
-                event = getEvent(accessToken, event.id()).event();
+                event = talkCalendarClientService.getEvent(accessToken, event.id()).event();
                 updatedEvents.add(event);
             } else {
                 updatedEvents.add(event);
@@ -66,28 +52,7 @@ public class TalkCalendarService {
 
     @Cacheable(value = "eventDetailCache", key = "#eventId")
     public EventDetailGetResponse getEvent(String accessToken, String eventId) {
-        String url = "https://kapi.kakao.com/v2/api/calendar/event";
-        URI uri = buildGetEventUri(url, eventId);
-
-        try {
-            ResponseEntity<EventDetailGetResponse> response = restClient.get()
-                .uri(uri)
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .retrieve()
-                .toEntity(EventDetailGetResponse.class);
-            return response.getBody();
-        } catch (Exception e) {
-            throw new RuntimeException("API 호출 중 오류 발생", e);
-        }
-    }
-
-    private URI buildUri(String baseUrl, String from, String to) {
-        return URI.create(
-            String.format("%s?from=%s&to=%s&limit=100&time_zone=Asia/Seoul", baseUrl, from, to));
-    }
-
-    private URI buildGetEventUri(String baseUrl, String eventId) {
-        return URI.create(String.format("%s?event_id=%s", baseUrl, eventId));
+        return talkCalendarClientService.getEvent(accessToken, eventId);
     }
 
     @CacheEvict(value = "eventListCache", key = "#memberId")
@@ -105,18 +70,8 @@ public class TalkCalendarService {
 
             // event라는 key로 JSON 데이터를 추가
             formData.add("event", eventJson);
-
-            // POST 요청 실행
-            ResponseEntity<Map> response = restClient.post()
-                .uri(uri)
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .header(HttpHeaders.CONTENT_TYPE, "application/x-www-form-urlencoded")
-                .body(formData)
-                .retrieve()
-                .toEntity(Map.class);
-
             // 응답에서 event_id 추출
-            Map<String, Object> responseBody = response.getBody();
+            Map<String, Object> responseBody =talkCalendarClientService.createEvent(accessToken, formData);
             if (responseBody != null && responseBody.containsKey("event_id")) {
                 String eventId = responseBody.get("event_id").toString();
                 // EventCreateResponseDto로 응답 반환
